requirements.txt
APScheduler==3.0.3
beautifulsoup4==4.4.0
boto3==1.1.3
botocore==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-toolbelt==0.0.1
docutils==0.12
futures==2.2.0
gunicorn==19.3.0
hashlib==20081119
hmac==20101005
jmespath==0.7.1
lxml==3.4.4
poster==0.8.1
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
six==1.9.0
static3==0.6.1
tzlocal==1.2
Unirest==1.1.7
wheel==0.24.0
